refactor(python): split python.rs into a directory module; extract arrays.rs
First step of the python.rs decomposition (Round 2). Converts the 19k-line
`src/python.rs` into a directory module `src/python/mod.rs` and extracts the
self-contained ndarray<->state adapters — `Arr2`/`Arr3`/`Arr3f32` and the
`arr*_opt`/`arr*_back` helpers — into `src/python/arrays.rs` (pub(crate),
glob-imported by mod.rs). Proves the directory-module pattern with the most
cohesive, lowest-risk helper group; error/save helpers and the per-model
families follow in subsequent PRs.

Pure move, no behavior change: public Python names unchanged, save/load round-
trips verified. cargo clippy -D warnings + fmt --check clean, cargo test
--workspace (133 + 16), maturin --features python, LDA/CTM save-load smoke.
